Feat: Component 분리-TeamSelect/Detail
TeamSelect/DetailPage에서 컴포넌트 분리 및 코드 리팩토링
-Gap, TeamAddModal같이 teamRouter안에서만 쓰일 것 같은 Component들은 일단 page 안의 components 파일에 넣어놨습니다.
- 전체적인 코드 리팩토링
- unitInfoRow button이 오른쪽으로 쏠려있는 문제 수정
