+++ src/pages/Team/Detail/TeamDetailPage.jsx
@@ -5,6 +5,9 @@ import {
   MEMBER_RAW_DATA_BASKETBALL,
 } from '../../../utils/formatData';
 import UnitInfoRow from '../../../components/ui/UnitInfoRow/UnitInfoRow';
+import Level from '../../../components/ui/Level/Level';
+import Gap from '../components/Gap';
+
 import * as S from './TeamDetailPage.style';
 
 const TEAM_INFO = {
@@ -18,8 +21,9 @@ const TEAM_INFO = {
 function TeamDetailPage() {
   const navigate = useNavigate();
 
+  //#TODO
   const renderMember = () =>
-    formatMemberData(MEMBER_RAW_DATA_BASKETBALL).map(member => (
+    formatMemberData(MEMBER_RAW_DATA_BASKETBALL).map((member) => (
       <UnitInfoRow
         key={member.id}
         unitInfo={member.unitInfo}
@@ -33,34 +37,21 @@ function TeamDetailPage() {
       <S.TeamBanner>
         <S.TeamLogo />
       </S.TeamBanner>
+
       <S.TeamNameSection>
         <S.TeamName>{TEAM_INFO.name}</S.TeamName>
-
-        <S.statusDiv>
-          <S.levelDiv>
-            <S.levelSpan>팀의 실력레벨</S.levelSpan>
-            <S.levelGauge>실력레벨</S.levelGauge>
-          </S.levelDiv>
-          <S.levelDiv>
-            <S.levelSpan>팀의 메너레벨</S.levelSpan>
-            <S.levelGauge>메너레벨</S.levelGauge>
-          </S.levelDiv>
-        </S.statusDiv>
-
+        <Level />
         <S.description>{TEAM_INFO.description}</S.description>
       </S.TeamNameSection>
-      <S.Gap />
+      <Gap />
 
       <S.TeamMembersSection>{renderMember()}</S.TeamMembersSection>
-      <S.GapWithSaveButton>
-        <S.SaveButton
-          onClick={() => {
-            navigate('/team/update');
-          }}
-        >
+
+      <Gap height='7.19rem'>
+        <S.SaveButton onClick={() => navigate('/team/update')}>
           수정하기
         </S.SaveButton>
-      </S.GapWithSaveButton>
+      </Gap>
     </S.Wrapper>
   );
 }
