+++ src/pages/Team/Detail/TeamDetailPage.style.jsx
@@ -9,12 +9,6 @@ export const Wrapper = styled.div`
 
   box-sizing: border-box;
 `;
-export const Gap = styled.div`
-  width: 100%;
-  height: 1.25rem;
-
-  background-color: var(--Blue100, #f0f0f0);
-`;
 
 export const TeamBanner = styled(Banner)``;
 export const TeamLogo = styled(CircleLogo)`
@@ -38,8 +32,6 @@ export const TeamNameSection = styled.section`
   padding: 1.5rem 8.33%;
 `;
 export const TeamName = styled.h2`
-  margin-bottom: 0.94rem;
-
   color: var(--Blue300, #0075ff);
   font-family: Pretendard;
   font-size: 28px;
@@ -49,49 +41,6 @@ export const TeamName = styled.h2`
   letter-spacing: -1.4px;
 `;
 
-export const statusDiv = styled.div`
-  display: flex;
-  justify-content: start;
-  align-items: center;
-
-  width: 100%;
-  margin-bottom: 1.63rem;
-`;
-export const levelDiv = styled.div`
-  display: flex;
-  justify-content: start;
-  align-items: center;
-`;
-export const levelSpan = styled.span`
-  margin-right: 0.38rem;
-
-  color: var(--kakao-logo, #000);
-  font-family: 'Pretendard Variable';
-  font-size: 12px;
-  font-style: normal;
-  font-weight: 400;
-  line-height: 20px; /* 166.667% */
-  letter-spacing: -0.6px;
-`;
-export const levelGauge = styled.p`
-  display: inline-flex;
-  justify-content: center;
-  align-items: center;
-
-  margin-right: 1.19rem;
-  padding: 0 6px;
-  gap: 10px;
-  border-radius: 3px;
-
-  background: var(--Blue200, #bddbff);
-  color: var(--Blue300, #0075ff);
-  font-family: 'Pretendard Variable';
-  font-size: 9px;
-  font-style: normal;
-  font-weight: 400;
-  line-height: 20px; /* 222.222% */
-  letter-spacing: -0.45px;
-`;
 export const description = styled.p`
   display: flex;
   justify-content: center;
@@ -99,6 +48,7 @@ export const description = styled.p`
 
   width: 94.5%;
   height: 3.5rem;
+  margin-top: 0.5rem;
   margin-left: auto;
   margin-right: auto;
   padding: 10px;
@@ -126,11 +76,6 @@ export const TeamMembersSection = styled.section`
   padding: 1.5rem;
 `;
 
-export const GapWithSaveButton = styled(Gap)`
-  position: relative;
-
-  height: 7.19rem;
-`;
 export const SaveButton = styled(Button300)`
   position: absolute;
   top: 25%;
