+++ src/pages/Team/Home/TeamSelectionPage.style.jsx
@@ -1,6 +1,5 @@
 import styled from 'styled-components';
 
-import { ModalButtonBlue } from '../../../components/ui/Modal/Modal';
 import CircleLogo from '../../../components/atoms/CircleLogo';
 
 // Style Content
@@ -10,67 +9,6 @@ export const Wrapper = styled.div`
   width: 100%;
 
   box-sizing: border-box;
-`;
-
-// Sport Selection
-export const SportSelectionSection = styled.section`
-  display: inline-flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: flex-start;
-
-  width: 100%;
-  padding: 5% 10%;
-  gap: 0.1875rem;
-
-  background: var(--white, #fff);
-
-  color: var(--Blue300, #0075ff);
-  /* 12md */
-  font-family: Pretendard;
-  font-size: 0.75rem;
-  font-style: normal;
-  font-weight: 500;
-  line-height: 1.375rem; /* 183.333% */
-  letter-spacing: -0.0375rem;
-
-  box-sizing: border-box;
-`;
-export const SportSelect = styled.select`
-  display: flex;
-  justify-content: center;
-  align-items: center;
-
-  width: 100%;
-  height: 2.5rem;
-  padding: 0.5rem 0.625rem;
-  gap: 14.5rem;
-  border-radius: 0.375rem;
-  border: 2px solid var(--Blue300, #0075ff);
-
-  background: url('/assets/chevron-down-5 1.svg') no-repeat 98% 50%/1.5rem auto;
-
-  color: var(--Blue300, #0075ff);
-  font-size: 16px;
-  font-weight: 500;
-
-  box-sizing: border-box;
-  -webkit-appearance: none; /* for chrome */
-  -moz-appearance: none; /*for firefox*/
-  appearance: none;
-
-  select::-ms-expand {
-    display: none; /*for IE10,11*/
-  }
-
-  &:focus {
-    //#TODO: 색깔 안변하는 문제
-    border: 2px solid var(--Blue300, #0075ff);
-  }
-
-  option {
-    border: 2px solid var(--Blue300, #0075ff);
-  }
 `;
 
 // Team Selection
@@ -83,9 +21,9 @@ export const TeamDiv = styled.div`
 
   background-size: cover;
   background-position: center;
-  background-color: ${props =>
+  background-color: ${(props) =>
     props.$backgroundColor || 'var(--Blue300, #0075ff)'};
-  color: ${props =>
+  color: ${(props) =>
     props.$backgroundColor === '#86ff91' ? '#0075ff' : '#FFF'};
 `;
 export const TeamName = styled.h1`
@@ -141,9 +79,4 @@ export const AddText = styled(TeamName)`
   top: 65%;
   left: 50%;
   transform: translateX(-50%);
-`;
-
-// for modal styling
-export const ConfirmBtn = styled(ModalButtonBlue)`
-  justify-content: center;
 `;
