+++ src/pages/Team/components/TeamAddModal/TeamAddModal.style.jsx
@@ -1,0 +1,6 @@
+import styled from 'styled-components';
+import { ModalButtonBlue } from '../../../../components/ui/Modal/Modal.style';
+
+export const ConfirmBtn = styled(ModalButtonBlue)`
+  justify-content: center;
+`;
